requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
watchdog
python-dotenv
construct==2.8.16
mgz==1.8.27
-e ./mgz_hd
flask-migrate
